requirements.txt
dlib==19.24.0
imutils==0.5.4
tensorflow==2.10.0
keras==2.10.0
lime==0.2.0.1
matplotlib==3.5.3
numpy==1.21.6
opencv-python==4.5.5
Pillow==9.2.0
scikit_image==0.19.3
scikit_learn==1.0.2
scipy==1.7.3
tqdm==4.66.1
